+++ particle_system/particle/Simulator.h
@@ -6,7 +6,9 @@
 
 #include "constants.h"
 
-typedef Eigen::Triplet<double> T;
+#include "Solver.h"
+
+typedef Eigen::Triplet<float> T;
 
 class Simulator
 {
@@ -17,7 +19,7 @@ public:
 
   void update();
 
-  std::array<double, SIZE>& getDensity() {
+  std::array<float, SIZE>& getDensity() {
     return m_density;
   }
 
@@ -38,30 +40,30 @@ private:
   double &m_time;
   
   // density
-  std::array<double, SIZE> m_density;
-  std::array<double, SIZE> m_density0;
+  std::array<float, SIZE> m_density;
+  std::array<float, SIZE> m_density0;
   // temperature
-  std::array<double, SIZE> m_temperature;
-  std::array<double, SIZE> m_temperature0;
+  std::array<float, SIZE> m_temperature;
+  std::array<float, SIZE> m_temperature0;
   // pressure
-  std::array<double, SIZE> m_pressure;
+  std::array<float, SIZE> m_pressure;
   // velocity
-  std::array<double, (Nx+1) * Ny * Nz> m_u,m_u0;
-  std::array<double, Nx * (Ny+1) * Nz> m_v,m_v0;
-  std::array<double, Nx * Ny * (Nz+1)> m_w,m_w0;
+  std::array<float, (Nx+1) * Ny * Nz> m_u,m_u0;
+  std::array<float, Nx * (Ny+1) * Nz> m_v,m_v0;
+  std::array<float, Nx * Ny * (Nz+1)> m_w,m_w0;
 
-  std::array<double, SIZE> m_avg_u, m_avg_v, m_avg_w;
+  std::array<float, SIZE> m_avg_u, m_avg_v, m_avg_w;
   // vorticity
-  std::array<double, SIZE> m_omg_x, m_omg_y, m_omg_z;
+  std::array<float, SIZE> m_omg_x, m_omg_y, m_omg_z;
   // force
-  std::array<double, SIZE> m_fx, m_fy, m_fz;
+  std::array<float, SIZE> m_fx, m_fy, m_fz;
 
 
   // solver
   std::vector<T> tripletList;
-  Eigen::ConjugateGradient<Eigen::SparseMatrix<double>, Eigen::Lower | Eigen::Upper> ICCG;
+  Eigen::ConjugateGradient<Eigen::SparseMatrix<float>, Eigen::Lower | Eigen::Upper> ICCG;
 
-  Eigen::SparseMatrix<double, Eigen::RowMajor> A;
-  Eigen::VectorXd b;
-  Eigen::VectorXd x;
+  Eigen::SparseMatrix<float, Eigen::RowMajor> A;
+  Eigen::VectorXf b;
+  Eigen::VectorXf x;
 };
